requirements.txt
numpy
scipy
matplotlib
sklearn
opencv-python
tensorflow-gpu==2.3.0
tensorflow-probability==0.11.1
